Fixed #45: custom messages for custom validators
I added a test for the 'allow the use of custom messages on failed validation', too.

Unfortunately the test-setup seems not to work well with the promise-based interface of Checkit. Even failing tests are counted as passed and you only see the errors by scanning the test-list for thrown errors. I added a `done`-callback in a final `then(done, done)` to the Checkit-Promise of my test to make it work. All other tests should be updated, too. But I'll open an issue for that.

There seemed to be some un-built changes of former commits. So lines 718 + 897 of `dist/checkit.js` are unrelated to this fix and contain new messages for the field-type *integer*.